+++ CHANGELOG.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Toplevel compile-time constants
+  ```
+  namespace N =
+    let nc = 1
+  contract C =
+    let cc = 2
+  ```
 ### Changed
 ### Removed
 ### Fixed
+- Warning about unused include when there is no include.
 
 ## [7.1.0]
 ### Added